W428: 4 new MATCH, nd -1437, control-flow shape is the tier's defect
First-party 6,867 -> 6,871 MATCH of 8,235 (83.4%); NONMATCHING 1,367 -> 1,363.
Total MATCH 8,033 -> 8,037. Normalized diff 1,237,299 -> 1,235,862, a drop of
1,437 - the largest of the session. 28 functions improved, one collateral +3,
zero MATCH lost, zero blocking statuses, no function pushed past its window.

New MATCH
- 003403b0 mdlEffect.c   nd  70 -> 0, 580/592   swap of two declarations
- 00348f30 mdlEffect.c   nd 115 -> 0, 344/352   float evaluation order plus an
                                                operand-order helper and an
                                                _sda global alias
- 003ed5e0 fclShop.c     nd 130 -> 0, 1952/1952 outer branch and body layout
                                                inverted
- 00452f70 y_misc.c      nd  67 -> 0, 864/864   register-role split of a
                                                post-scan loop plus direction
                                                temporaries and a commutative
                                                helper

Target selection
----------------
Rather than grinding the nd<=60 band again, this wave took the 86 functions with
nd between 61 and 150 whose object is within 16 bytes of its retail window. That
combination isolates pure CODEGEN SHAPE: the code is all present, so nothing is
missing and nothing needs decoding. The tier had never been worked.

The finding: this tier is control-flow shape, not colouring
-----------------------------------------------------------
Register-role colouring explains nd 1-20. At nd 60-150 the residual is too large
to be two transposed registers, and the dominant defect is the SHAPE of the
control flow - which statements sit inside a branch, whether a loop is rotated,
where a body is placed relative to its test. This was not in the wave's original
brief; it emerged from the first conversion and was broadcast mid-wave, after
which most of the remaining results came from it.

What worked, with examples:
- Invert an if/else so the body retail places out-of-line sits behind the taken
  branch. 003ed5e0 nd130 -> 0. 002bce10 nd130 -> 76. 00110510 nd106 -> 1.
- Restructure a loop into a `for` with an explicit `break`. 003da570 nd138 -> 36.
- Express a guard as an explicit taken-skip `goto`. Used in 002bce10 and
  00110510.
- Guard a call that retail only makes conditionally. 00361350 nd119 -> 9.

Largest improvements without conversion
- 003c2ee0 comuSasoi.c    nd 130 -> 4    return-width correction, declaration
                                          order, branch-body layout
- 00110510 h_maestro.c    nd 106 -> 1    explicit goto/label layout
- 00361350 mt_evtMain.c   nd 119 -> 9    guarded call plus u32 stack-field width
- 003f7730 fclShopMisc.c  nd 106 -> 8    typed float calls, signed text
                                          argument, final-call hoist
- 001fc980 pb_main.c      nd 142 -> 17   third call parameter u16 -> s16, and a
                                          u16 return; retail uses dsll32/dsra32
                                          0x10 and andi 0xffff
- 003b7090 mt_scene.c     nd  97 -> 25   absolute global aliases plus
                                          declaration order
- 0032a770 mdlEffect.c    nd 137 -> 66
- 003da570 fclCombineList nd 138 -> 36
- 002bce10 btlEffect.c    nd 130 -> 76
- 001365b0 h_camp.c       nd 108 -> 90
- 003f5ab0 / 003f5d10 fclShopMisc, both nd 109 -> 88 (twins, same fix)

A parameter permutation that improves a callee can break its callers
--------------------------------------------------------------------
nmSequence's 003c5d40 improved nd130 -> 89 under a parameter-order permutation,
but changing a definition's parameter order changes its ABI, and the two callers
in fclShopMisc went from MATCH to MISMATCH. Retail's caller at 00409f50 sets up
`move a0,$zero; lw a1,4($s3)` - the ORIGINAL order - so the permutation was
simply wrong. Reverted on both sides; both callers are back at nd0. Any family-2
change to a definition must be verified across every file that calls it.

Floors closed with evidence
- cosf, sinf and tanf all begin `lui $t5,0x7fff; mfc1 $t7,$f12; ori $t5,0xffff`
  and mask. MWCC has no path to `mfc1`; every bit-reinterpretation form tested
  emits stack `swc1`/`lw` instead. This is the same ee-gcc provenance already
  established for sf_fabs and the nine 24-byte sf_tan wrappers.
- scrComu 00360ed0: retail's case-0 branch reuses `v0=1` left by the previous
  case's compare, while MWCC always emits `addiu v0,1`. The extra 4 bytes shift
  everything after it. No semantics-preserving source shape suppresses it.
- mdlEffect 003269e0: parameters and locals have exchanged register banks
  (params in s3/s2 where retail has s1/s0). Tail-ordering and parameter-copying
  both measured flat, and the widths match, so it is a colouring floor.
- k_footstep, k_sceneDraw, k_data: every probe rejected, files restored.

Rejected under the acceptance rule
`#pragma schedule on` improved uty_tmr 7fe80's nd but shrank the object 20 bytes
away from its window and worsened the rate 0.667 -> 0.689. Optimisation-pragma
size changes are banned; reverted.

Verified: the build links byte-identically (SLUS_216.21 sha1
3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb OK), verify.py scans the same 14,055
functions with no compile errors, no missing symbols and no unknown addresses,
all 92 tests pass, and a comment-aware check finds zero unbalanced pragma files.

+++ src/Main/Social/pb_main.c
@@ -245,9 +245,9 @@ int func_001fc870(u8* event, void* target)
 
 void func_001fc980(u8* event, u16* result, s32* count)
 {
-    extern void FUN_0017be10(int param_1, int param_2, u16 param_3, int param_4,
+    extern void FUN_0017be10(int param_1, int param_2, s16 param_3, int param_4,
                              void* param_5, void* param_6, void* param_7);
-    extern u8 FUN_0017bb40(s32 id);
+    extern u16 FUN_0017bb40(s32 id);
     DatPersonaWork* persona;
     u16* skills;
     u32 skillCount;
